+++ reify_test.go
@@ -20,6 +20,7 @@ package ucfg
 import (
 	"testing"
 
+	"github.com/elastic/go-ucfg/parse"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -115,6 +116,7 @@ func TestUnpackPrimitivesValuesResolve(t *testing.T) {
 			U uint
 			F float64
 			S string
+			W string
 		}{},
 		&struct {
 			B interface{}
@@ -122,6 +124,7 @@ func TestUnpackPrimitivesValuesResolve(t *testing.T) {
 			U interface{}
 			F interface{}
 			S interface{}
+			W interface{}
 		}{},
 		&struct {
 			B *bool
@@ -129,19 +132,21 @@ func TestUnpackPrimitivesValuesResolve(t *testing.T) {
 			U *uint
 			F *float64
 			S *string
+			W *string
 		}{},
 	}
 
 	cfgOpts := []Option{
 		VarExp,
-		Resolve(func(name string) (string, error) {
+		Resolve(func(name string) (string, parse.Config, error) {
 			return map[string]string{
 				"v_b": "true",
 				"v_i": "42",
 				"v_u": "23",
 				"v_f": "3.14",
 				"v_s": "string",
-			}[name], nil
+				"v_w": "{string}",
+			}[name], parse.EnvConfig, nil
 		}),
 	}
 
@@ -151,6 +156,7 @@ func TestUnpackPrimitivesValuesResolve(t *testing.T) {
 		"u": "${v_u}",
 		"f": "${v_f}",
 		"s": "${v_s}",
+		"w": "${v_w}",
 	}, cfgOpts...)
 
 	for i, out := range tests {
@@ -186,11 +192,15 @@ func TestUnpackPrimitivesValuesResolve(t *testing.T) {
 		s, err := c.String("s", -1, cfgOpts...)
 		assert.NoError(t, err)
 
+		w, err := c.String("w", -1, cfgOpts...)
+		assert.NoError(t, err)
+
 		assert.Equal(t, true, b)
 		assert.Equal(t, 42, int(i))
 		assert.Equal(t, 23, int(u))
 		assert.Equal(t, 3.14, f)
 		assert.Equal(t, "string", s)
+		assert.Equal(t, "{string}", w)
 	}
 }